% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runIntegrho.R
\name{runIntegrho}
\alias{runIntegrho}
\title{runIntegrho}
\usage{
runIntegrho(host = "127.0.0.1", port = 3277)
}
\value{
none
}
\description{
executes integrho app in default browser
}
\examples{
runIntegrho(host="127.0.0.1", port=3277)

}
